Temporary move trimIndent implementations to tests as private members to avoid resolution ambiguity
